backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
tensorflow==2.16.1
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
opencv-python==4.8.1.78
scikit-learn==1.3.2
pandas==2.1.4
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.4
redis==5.0.1
whitenoise==6.6.0
gunicorn==21.2.0
